requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-openai
faiss-cpu
pypdf
requests